requirements.txt
# OpenAI GPT Chatbot - Requirements
# Author: RSK World (https://rskworld.in)
# Email: help@rskworld.in
# Phone: +91 93305 39277
# Year: 2026

openai>=1.0.0
python-dotenv>=1.0.0
flask>=2.3.0

